Update dependencies
Version 1.7.0 of JSCS just came out with the `--esnext` option for ES6
compatibility. When running jsxcs with `--esnext`, I get the following
error:

> jsxcs --esnext path/to/jsx
> error: unknown option `--esnext'

I think this would be fixed by updating the JSCS dependency version.

While I was at it, I updated the versions of the other dependencies that
have newer versions, and sorted the list of dependencies alphabetically.

Fixes #1. Fixes #3.